requirements.txt
flask
flask-cors
python-dotenv
flask-sqlalchemy
flask-jwt-extended
gunicorn
